# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      err_code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      err_code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
...
  HTTP500InternalServerError:
    properties:
      err_code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.CCVMCreateRequest:
    properties:
      cpu:
        description: vCPU 개수
        example: 4
        type: integer
      gfs_mount_point:
        description: GFS mount point. ablestack-vm에서 ccvm.qcow2 경로 계산에 사용한다.
        example: /mnt/glue-gfs
        type: string
      management_network_bridge:
        description: management network bridge
        example: br0
        type: string
      memory:
        description: 메모리 크기(GiB)
        example: 16
        type: integer
      service_network_bridge:
        description: optional service network bridge
        example: br1
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.CCVMCreateResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: ccvm create success
        type: string
      setup:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMPCSControlResponse'
      xml:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMXMLCreateResponse'
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.CCVMPCSControlResponse:
    properties:
      action:
        example: status
        type: string
      code:
        example: 200
        type: integer
      message:
        type: string
      retname:
        type: string
      target:
        example: 10.10.31.1
        type: string
      val: {}
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.CCVMXMLCreateResponse:
    properties:
      code:
        description: 처리 결과 코드
        example: 200
        type: integer
      message:
        description: 처리 결과 메시지
        example: 클라우드센터 가상머신 xml 생성 성공
        type: string
      results:
        description: fan-out 대상별 결과
        items:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyResult'
        type: array
      val:
        description: 기존 Python createReturn 호환 결과값
      xml_path:
        description: 생성된 XML 파일 경로
        example: /etc/ablestack/vmconfig/ccvm/ccvm.xml
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyLocalResponse:
    properties:
      code:
        example: 200
        type: integer
      val:
        example: Cluster Config insert Success
        type: string
    type: object
  ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyRequest:
    properties:
      action:
        description: 'action: insert/remove/reset/check'
        example: insert
        type: string
      ccvm:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterCCVMConfig'
        description: ccvm config (ip only)
      exclude_hostname:
        description: exclude hostname (hci only)
        example: ablecube31-2
        type: string
      external_timeserver:
        description: external timeserver
        example: time.google.com
        type: string
      hosts:
        description: hosts list (preferred)
        items:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterHost'
        type: array
      mngtNic:
        allOf:
        - $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterMngtNicConfig'
        description: management NIC config
      new_hostname:
        description: 'new hostname (option=add: exclude scvm fanout)'
        example: ablecube12-3
        type: string
      option:
        description: option (e.g. add/hostOnly/withScvm/all/local)
        example: add
        type: string
      pcs_cluster_list:
        description: pcs cluster node IP list
        example:
        - 10.10.31.1
        - 10.10.31.2
        - 10.10.31.3
        items:
          type: string
        type: array
      remove_hostname:
        description: remove hostname
        example: ablecube31-3
        type: string
      storage_network:
        description: storage network usage (true/false)
        example: "false"
        type: string
      type:
        description: cluster type (e.g. ablestack-vm, ablestack-hci)
        example: ablestack-vm
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyResult:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: ok
        type: string
      target:
        example: 10.10.31.1
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterCCVMConfig:
    properties:
      ip:
        description: ccvm IP
        example: 10.10.31.10
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterHost:
    properties:
      ablecube:
        example: 10.10.31.1
        type: string
      ablecubePn:
        example: 100.100.31.1
        type: string
      hostname:
        example: ablecube31-1
        type: string
      index:
        example: "1"
        type: string
      scvm:
        example: 100.100.31.11
        type: string
      scvmCn:
        example: 100.200.31.11
        type: string
      scvmMngt:
        example: 10.10.31.11
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.ClusterMngtNicConfig:
    properties:
      cidr:
        description: management NIC CIDR
        example: "16"
        type: string
      dns:
        description: management NIC DNS
        example: 8.8.8.8
        type: string
      gw:
        description: management NIC gateway
        example: 10.10.0.1
        type: string
    type: object
...
  ablecloud_io_ablestack-api_internal_model_cube.LicenseResponse:
    properties:
      code:
        example: 200
        type: integer
      val: {}
    type: object
...
  ablecloud_io_ablestack-api_internal_service_authservice.TokenClaim:
    properties:
      exp:
        type: integer
      iat:
        type: integer
      role:
        type: string
      sub:
        type: string
      use:
        type: string
    type: object
  internal_handler_auth.InternalTokenApplyRequest:
    properties:
      internal_token:
        type: string
    type: object
  internal_handler_auth.InternalTokenResult:
    properties:
      code:
        type: integer
      message:
        type: string
      target:
        type: string
    type: object
  internal_handler_auth.InternalTokenRotateResponse:
    properties:
      code:
        example: 200
        type: integer
      message:
        type: string
      results:
        items:
          $ref: '#/definitions/internal_handler_auth.InternalTokenResult'
        type: array
      val:
        additionalProperties: {}
        type: object
    type: object
  internal_handler_auth.LoginRequest:
    properties:
      id:
        example: root
        type: string
      password:
        example: password
        type: string
    type: object
  internal_handler_auth.LoginResponse:
    properties:
      access_token:
        type: string
      authorization:
        example: Bearer eyJ...
        type: string
      code:
        example: 200
        type: integer
      expires_in:
        example: 3600
        type: integer
      token_type:
        example: Bearer
        type: string
    type: object
  internal_handler_auth.MeResponse:
    properties:
      code:
        example: 200
        type: integer
      val:
        $ref: '#/definitions/ablecloud_io_ablestack-api_internal_service_authservice.TokenClaim'
    type: object
...
info:
  contact:
    email: ycyun@ablecloud.io
    name: API Support
    url: https://www.ablecloud.io/support
  description: This is a Cube-API server.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://ablecloud.io/
  title: Cube API
  version: "1.0"
paths:
  /auth/internal-token/apply:
    post:
      security: []
      consumes:
      - application/json
      description: 내부 API 서버 간 호출로 전달받은 X-Cube-Internal-Token 값을 현재 호스트에 적용합니다.
      parameters:
      - description: current internal token
        in: header
        name: X-Cube-Internal-Token
        required: true
        type: string
      - description: internal token apply request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_auth.InternalTokenApplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Apply Internal Token
      tags:
      - Auth
  /auth/internal-token/rotate:
    post:
      description: 클러스터 내부 API 호출용 X-Cube-Internal-Token 값을 교체하고 대상 AbleCube 노드에 적용합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_auth.InternalTokenRotateResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Rotate Internal Token
      tags:
      - Auth
  /auth/login:
    post:
      security: []
      consumes:
      - application/json
      description: 웹 UI/API용 access token을 발급합니다.
      parameters:
      - description: login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_auth.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - Auth
  /auth/me:
    get:
      description: Bearer token을 검증하고 현재 token claim을 반환합니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler_auth.MeResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Current User
      tags:
      - Auth
...
  /cube/ccvm/create:
    post:
      consumes:
      - application/json
      description: 제품 타입에 맞는 CCVM XML과 backing image를 준비하고 PCS 리소스를 비활성 상태로 생성합니다.
        시작은 /cube/ccvm/lifecycle action=start를 사용합니다.
      parameters:
      - description: ccvm create request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.CCVMCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Create CCVM
      tags:
      - Cube-CCVM
...
  /cube/cluster/apply-local:
    post:
      consumes:
      - application/json
      description: 로컬 노드에서만 cluster_config CLI를 실행합니다. insert 적용 시 시간 서버 설정도 함께 적용합니다.
      parameters:
      - description: apply request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.ClusterApplyLocalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Apply Cluster Config (Local)
      tags:
      - Cube-Cluster
...
  /cube/license:
    post:
      security: []
      consumes:
      - multipart/form-data
      description: 라이센스 등록 및 상태 확인을 수행합니다. 등록은 JSON license_content 또는 multipart license_file/file
        업로드를 지원합니다.
      parameters:
      - description: status/register
        in: formData
        name: action
        required: true
        type: string
      - description: license file
        in: formData
        name: license_file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ablecloud_io_ablestack-api_internal_model_cube.LicenseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: License Control
      tags:
      - Cube-License
...
schemes:
- http
- https
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Enter the token with the Bearer prefix, e.g. Bearer eyJ...
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"